requirements.txt
openai==0.27.8
streamlit==1.25.0
python-dotenv==0.21.1
psycopg2==2.9.7
thefuzz==0.19.0
